Refactor Winning Player

+++ pong-sdl/pong-sdl/pixel.h
@@ -2,6 +2,7 @@
 #define __PONG_SDL_PIXEL__
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 
 #include "./constants.h"
@@ -10,10 +11,10 @@ enum EBlockTypes { BACKGROUND_BLOCK = 0b0, NET_BLOCK = 0b1, PLAYER_BLOCK = 0b10,
 
 typedef struct _pixel Pixel;
 
-void setPixelX(Pixel*, size_t);
-void setPixelY(Pixel*, size_t);
-size_t getPixelX(Pixel*);
-size_t getPixelY(Pixel*);
+void setPixelX(Pixel*, int32_t);
+void setPixelY(Pixel*, int32_t);
+int32_t getPixelX(Pixel*);
+int32_t getPixelY(Pixel*);
 
 Pixel* initPixel(size_t x, size_t y, enum EBlockTypes blockType);
 void destroyPixel(Pixel* pixel);
